assignment_RAG/requirements.txt
langchain==0.3.20
langgraph==0.3.5
openai==1.65.4
pinecone==6.0.1
pypdf==5.3.1
python-dotenv==1.0.1
